changed two muon-id variables from double to int
git-svn-id: file:///home/home1/institut_3a/chof/Rep/MIS/trunk/ePaxDemo@252 4fa44358-d421-0410-8b48-b5a29ab96256

+++ ePaxAnalyzer/src/ePaxAnalyzer.cc
@@ -755,10 +755,11 @@ void ePaxAnalyzer::analyzeRecMuons(const edm::Event& iEvent, pxl::EventViewRef E
 
 	 //save some stuff related to Muon-ID (Calo-info etc.)
 	 part.set().setUserRecord<double>("CaloCompatibility", muon->getCaloCompatibility());
-	 part.set().setUserRecord<double>("NumberOfChambers", muon->numberOfChambers());
-	 part.set().setUserRecord<double>("NumberOfMatches", muon->numberOfMatches());
+	 part.set().setUserRecord<int>("NumberOfChambers", muon->numberOfChambers());
+	 part.set().setUserRecord<int>("NumberOfMatches", muon->numberOfMatches());
 	 part.set().setUserRecord<double>("MuonDepositEM", muon->getCalEnergy().em);
 	 part.set().setUserRecord<double>("MuonDepositHCAL", muon->getCalEnergy().had);
+	
 	
          numMuonRec++;
       }
